# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ARCHIVE API (native clients)
  version: 1.0.0
  description: |
    The contract archive-ios and archive-android are built against. Every path is a Next.js route
    under `app/api`; the loaders behind the GET bundles are the same functions the web pages
    render from (`lib/loaders/*`), so a screen on the app and its page on the web see identical
    data.

    **Auth.** Send `Authorization: Bearer <supabase access token>` on every authenticated call.
    The token is validated server-side per request. Refresh it client-side (supabase-swift) before
    it expires; an expired token yields `401`. Also send `X-Client: ios/<build>` so the server can
    log the caller and apply the force-upgrade floor from `GET /api/mobile/config`.

    **Errors.** `{ "error": string, "code"?: string }`. 400 malformed input · 401 no/invalid session ·
    403 forbidden (`code: banned`, `verification_required`) · 404 not found **or feature flag off** ·
    409 conflict / lock / window closed · 422 semantic rejection · 429 rate limited (`Retry-After`
    seconds) · 502 payment/SMS provider error · 503 not configured.

    **Compatibility policy.** Additive changes only on existing paths. A breaking change ships on
    a new path and bumps `min_build` in `GET /api/mobile/config`.

    **Money.** Integer cents everywhere. Clients display only what the server returns; they never
    compute fees or totals.

...

...
    to the seller/admin.
  x-repo: aetiusgular/resale-platform
servers:
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: No or invalid session.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: Unauthorized } } }
    NotFound:
      description: Not found, or the feature flag is off.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' }, example: { error: Not found } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string, description: 'Stable machine code when present: banned, verification_required, auth_required, invalid_username, phone_in_use, prohibited_block, unconfigured, username_collision.' }
...
    UserSizes:
      type: object
      description: '`{ "<dept>:<section>": [sizes] }`, e.g. `{"menswear:tops":["M","L"]}`.'
      additionalProperties: { type: array, items: { type: string } }
...
    LcTally:
      type: object
      properties:
        legit: { type: integer }
        flagged: { type: integer }
        autoAuth: { type: string, enum: [TAG PASS, TAG FAIL, UNCERTAIN, PENDING] }
        verdict: { type: string, enum: [LEGIT, NOT LEGIT, PENDING] }
    ListingDetail:
      type: object
      required: [listing, seller, viewer, lc, flags]
      properties:
        listing:
          type: object
          properties:
            id: { type: string, format: uuid }
            title: { type: string }
            brand: { type: string }
            category: { type: string }
            department: { type: string }
            subcategory: { type: [string, 'null'] }
            size: { type: [string, 'null'] }
            color: { type: [string, 'null'] }
            description: { type: [string, 'null'] }
            condition_score: { type: [integer, 'null'] }
            condition_notes: { type: [string, 'null'] }
            measurements: { type: object, additionalProperties: { type: number }, description: 'Inches, keyed by the labels in measurement_labels.' }
            measurement_labels: { type: array, items: { type: string } }
            price_cents: { type: integer }
            price_display: { type: string }
            original_price_cents: { type: [integer, 'null'] }
            shipping_cents: { type: integer }
            saves_count: { type: integer }
            view_count: { type: integer }
            is_price_dropped: { type: boolean }
            images: { type: array, items: { type: string, format: uri }, description: 'Public slots 0–4.' }
            possession_photo_url: { type: [string, 'null'], description: 'Seller/admin only.' }
            status: { type: string, enum: [draft, pending_review, active, pending_escrow, sold, removed] }
            status_word: { type: string }
            rejection_reason: { type: [string, 'null'] }
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
            authentication_status: { type: string }
            authenticated: { type: boolean }
            listed_line: { type: string, example: 'LISTED 3D AGO · 18 SAVED' }
            spec: { type: string, example: 'M · BLACK' }
            crumb: { type: object, properties: { parts: { type: array, items: { type: string } }, href: { type: string } } }
        seller:
          type: object
          properties:
            id: { type: string, format: uuid }
            username: { type: string }
            initials: { type: string }
            role: { type: [string, 'null'] }
            id_verification_status: { type: [string, 'null'] }
            verified: { type: boolean }
            trust_line: { type: string, example: '4.9 · 132 SALES · VERIFIED ID' }
        viewer:
          type: object
          properties:
            id: { type: [string, 'null'], format: uuid }
            username: { type: string }
            is_seller: { type: boolean }
            is_admin: { type: boolean }
            saved: { type: boolean }
            following: { type: boolean }
            can_buy: { type: boolean }
            can_post: { type: boolean }
            can_edit: { type: boolean }
        lc: { $ref: '#/components/schemas/LcTally' }
        flags: { type: object, properties: { bump: { type: boolean }, boost: { type: boolean }, follows: { type: boolean }, auth_badge: { type: boolean } } }
...
    Address:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        street1: { type: string }
        street2: { type: [string, 'null'] }
        city: { type: string }
        state: { type: string, pattern: '^[A-Z]{2}$' }
        zip: { type: string, pattern: '^\d{5}(-\d{4})?$' }
        country: { type: string, const: US }
        is_default: { type: boolean }
        created_at: { type: string, format: date-time }
    NotificationPrefs:
      type: object
      description: 16 booleans, one email + one push per bucket (offers, offer_result, messages, sold, price_drops, search_alerts, orders, alerts). Defaults true.
      additionalProperties: { type: boolean }
    SideDashboard:
      type: object
      properties:
        side: { type: string, enum: [buyer, seller] }
        volumeCents: { type: integer }
        orderCount: { type: integer }
        activityBps: { type: integer }
        effectiveBps: { type: integer }
        locked: { type: boolean }
        lockedUntilMs: { type: [integer, 'null'] }
        current: { type: object }
        next: { type: [object, 'null'] }
        volumeToNextCents: { type: integer }
        ordersToNext: { type: integer }
        expiringVolumeCents: { type: integer }
        expiringOrderCount: { type: integer }
    SettingsOrderRow:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [buyer, seller] }
        state: { type: string }
        brand: { type: string }
        title: { type: string }
        size: { type: string }
        image: { type: [string, 'null'] }
        amount: { type: string, example: '$210' }
        counterparty: { type: string }
        created_at: { type: string, format: date-time }
        hasTracking: { type: boolean }
        canReview: { type: boolean }
    ReviewTarget:
      type: object
      properties:
        orderId: { type: string, format: uuid }
        brand: { type: string }
        title: { type: string }
        image: { type: [string, 'null'] }
        amount: { type: string }
        deliveredAt: { type: string, format: date-time }
        sellerUsername: { type: string }
        eligible: { type: boolean }
        reason: { type: [string, 'null'] }
    SettingsData:
      type: object
      description: The exact object /settings renders from. camelCase is historical (web-first); the app decodes it as-is.
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: [string, 'null'] }
        avatarUrl: { type: [string, 'null'] }
        usernameNextChangeAt: { type: [integer, 'null'], description: 'ms epoch' }
        usernameWindowOpen: { type: boolean }
        tierReviewDateMs: { type: integer }
        email: { type: string }
        emailVerified: { type: boolean }
        memberSince: { type: string, example: '08/2026' }
        idVerified: { type: boolean }
        sizes: { $ref: '#/components/schemas/UserSizes' }
        hideNotMySize: { type: boolean }
        addresses: { type: array, items: { $ref: '#/components/schemas/Address' } }
        payoutsEnabled: { type: boolean }
        payoutOnboardingDone: { type: boolean }
        prefs: { $ref: '#/components/schemas/NotificationPrefs' }
        notificationsEnabled: { type: boolean }
        phoneVerificationEnabled: { type: boolean }
        phoneVerified: { type: boolean }
        phone: { type: [string, 'null'] }
        tierDashboardEnabled: { type: boolean }
        buyerTier: { oneOf: [{ $ref: '#/components/schemas/SideDashboard' }, { type: 'null' }] }
        sellerTier: { $ref: '#/components/schemas/SideDashboard' }
        welcomeLeft: { type: integer }
        salesCount: { type: integer }
        reviewsEnabled: { type: boolean }
        shippingLabelsEnabled: { type: boolean }
        orders: { type: array, items: { $ref: '#/components/schemas/SettingsOrderRow' }, description: 'Empty here; use GET /api/settings/orders.' }
        review: { oneOf: [{ $ref: '#/components/schemas/ReviewTarget' }, { type: 'null' }] }
    SettingsBundle:
      type: object
      required: [data, activeOrders]
      properties:
        data: { $ref: '#/components/schemas/SettingsData' }
        activeOrders: { type: integer }
...
    ListingInitial:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        title: { type: [string, 'null'] }
        brand: { type: [string, 'null'] }
        category: { type: [string, 'null'] }
        department: { type: [string, 'null'] }
        subcategory: { type: [string, 'null'] }
        size: { type: [string, 'null'] }
        color: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        condition_score: { type: [integer, 'null'] }
        price_cents: { type: [integer, 'null'] }
        images: { type: array, items: { type: string } }
        possession_photo_url: { type: [string, 'null'] }
        measurements: { type: object, additionalProperties: { type: number } }
...
    BoostState:
      type: object
      properties:
        viewer: { type: object, properties: { username: { type: string } } }
        listing: { type: object, properties: { id: { type: string }, title: { type: string }, brand: { type: string }, active: { type: boolean }, boosted_until: { type: [string, 'null'] }, bumped_at: { type: [string, 'null'] } } }
        free_bump: { type: [object, 'null'], properties: { availableNow: { type: boolean }, nextAtIso: { type: [string, 'null'] } } }
        packages: { type: array, items: { type: object, properties: { key: { type: string }, label: { type: string }, amountCents: { type: integer }, durationDays: { type: integer } } } }
        flags: { type: object, properties: { boost: { type: boolean }, bump: { type: boolean } } }
...
security: []
paths:
...
  /api/listings/{id}:
    get:
      tags: [listings]
      summary: Listing detail bundle (public for active + sold; seller/admin otherwise).
      security: [{}, { bearer: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListingDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [selling]
      summary: 'Edit own listing (drafts: any field; live: title, description, price, size, colour, category tree, measurements, condition).'
      security: [{ bearer: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ListingInitial' } } } }
      responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { description: banned } }
    delete:
      tags: [selling]
      summary: Delete own draft.
      security: [{ bearer: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Unauthorized' } }
...
  /api/boosts:
    get:
      tags: [selling]
      summary: Boost page state for your own listing (flag BOOSTED_POSTS_ENABLED).
      security: [{ bearer: [] }]
      parameters: [{ name: listingId, in: query, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/BoostState' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [payments]
      summary: Buy a boost → PaymentIntent client secret for PaymentSheet.
      security: [{ bearer: [] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [listingId, package], properties: { listingId: { type: string, format: uuid }, package: { type: string, enum: [spotlight_3, feature_7, premier_14] } } } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { clientSecret: { type: string }, amountCents: { type: integer } } } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' } }

  # ── settings ─────────────────────────────────────────────────────────────────
  /api/settings/profile:
    get:
      tags: [settings]
      summary: Settings bundle (profile, prefs, addresses, tiers, derived windows).
      security: [{ bearer: [] }]
      parameters: [{ name: onboarding, in: query, schema: { type: string, enum: [complete] } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SettingsBundle' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [settings]
      summary: Update username (1×/30d), display name, avatar URL (avatars/{uid}/… on our storage).
      security: [{ bearer: [] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { username: { type: string }, display_name: { type: string }, avatar_url: { type: string, format: uri } } } } } }
      responses: { '200': { description: OK }, '400': { description: invalid_username }, '401': { $ref: '#/components/responses/Unauthorized' }, '409': { description: Username taken / window closed. } }
...